Proper handling of User on SIP URI
One of the added tests show how user would be incorrectly inferred because the original code supposed that everything until @ was the user part.
The other test shows a username with a period being considered a hostname (including the at symbol).
Changed how auth part of the SIP URI is parsed by using the characters as expressed by the SIP-URI part of the grammar on the section 25.1 of the SIP protocol RFC (RFC 3261).